implemented LoRAs inputs generation by backend
